Refactor: Extract inline CSS to classes and split App.tsx into custom hooks
- Moved solarized color palette from inline object to CSS variables
- Extracted inline styles for app container and creator credit to CSS classes
- Created useAudioSpeechRecognition hook to handle all audio/speech recognition logic
- Created useKeyboardShortcuts hook to handle keyboard event handling
- Significantly reduced App.tsx complexity while maintaining all functionality
- No breaking changes - all features work as before
