+++ frontend/src/App.css
@@ -26,6 +26,50 @@
   --solarized-base3: 253, 246, 227;
 }
 
+/* Solarized theme app styles */
+.app-solarized-dark {
+  background-color: rgb(var(--solarized-base03));
+  color: rgb(var(--solarized-base0));
+  min-height: 100vh;
+  transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.app-solarized-light {
+  background-color: rgb(var(--solarized-base3));
+  color: rgb(var(--solarized-base00));
+  min-height: 100vh;
+  transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+/* Creator credit styles */
+.creator-credit {
+  position: fixed;
+  bottom: 12px;
+  left: 12px;
+  font-size: 11px;
+  z-index: 50;
+  opacity: 0.7;
+  transition: opacity 0.3s ease;
+}
+
+.creator-credit:hover {
+  opacity: 1;
+}
+
+.creator-credit a {
+  color: inherit;
+  text-decoration: none;
+  font-family: system-ui, -apple-system, sans-serif;
+}
+
+.creator-credit-dark {
+  color: rgb(var(--solarized-base01));
+}
+
+.creator-credit-light {
+  color: rgb(var(--solarized-base1));
+}
+
 /* App container */
 .app {
   min-height: 100vh;
